requirements.txt
Django>=5.2
djangorestframework
django-cors-headers
django-filter
drf-yasg
Pillow
stripe
python-dotenv
pyotp
qrcode[pil]
imagehash
